IDE-908
Enabled layout editor to display nested layout.
